Add max_date to rss_activity.
Max_date is another quality of life function. It compares two datetimes, 
and is essentially a normal max that can handle Nones. This is useful 
since for each entity we can simply record the latest of all of the 
dates we encounter, and we no longer need to worry about handling Nones. 

Nones are always lower, this gives the behavior that entities without 
dates are treated as though they have not had any activity that is 
recent enough. This is also the same behavior that would occur if the 
entity was simply not present at all.