Add get_entry_last_modified to rss_activity.
The next few commits will be working up the heirarchy from finding the 
last active date from entries, to feeds, to companies.

Finally, after being able to determine the last active date of a 
company, the only remaining task will be to create a simple function 
which checks that last active date to determine whether or not it has 
been inactive for long enough.

This commit also makes a few a aesthetic changes, such as fixing the 
indentation and spacing between top level statements.